feat(typed-router): build commonjs and esm
