+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "license": "Apache-2.0",
   "scripts": {
-    "build": "nx build",
+    "build": "turbo build",
     "ensure": "node tools/scripts/ensure.mjs",
     "lint": "turbo lint",
     "release": "npm run build && npx changeset publish",
@@ -15,6 +15,7 @@
     "@google-cloud/pubsub": "3.7.1",
     "@swc/helpers": "0.5.1",
     "escape-goat": "4.0.0",
+    "npm-run-all": "^4.1.5",
     "tslib": "2.5.3"
   },
   "devDependencies": {
@@ -33,13 +34,12 @@
     "@typescript-eslint/parser": "5.60.0",
     "eslint": "8.43.0",
     "eslint-config-prettier": "8.8.0",
-    "jest": "29.5.0",
-    "jest-environment-node": "29.5.0",
     "nx": "16.4.0",
     "nx-cloud": "latest",
     "prettier": "2.8.8",
     "ts-jest": "29.1.0",
     "ts-node": "10.9.1",
+    "tsup": "7.1.0",
     "typescript": "5.1.3",
     "vitest": "^0.32.2"
   },
