--- packages/typed-router/src/lib/TypedRouter.spec.ts
@@ -1,5 +1,0 @@
-import { test, expect } from 'vitest'
-
-test('should ', () => {
-  expect(1).toBe(1)
-})
