+++ turbo.json
@@ -1,6 +1,7 @@
 {
   "$schema": "https://turbo.build/schema.json",
   "pipeline": {
+    "build": {},
     "lint": {},
     "test": {}
   }
